Move @Transactional from individual methods to class definition for UserIntegrationTest. This is needed to ensure tests don't persist any data until explicitly directed so by @Rollback(false) Move request mappings for individual operations to class definiton for UserController to avoid duplication. deleteUser now returns 400 with description if user cannot be found by id.

+++ src/main/java/com/github/adorogensky/workshop/alpha/service/UserService.java
@@ -87,6 +87,15 @@ public class UserService {
 	}
 
 	public void deleteUser(Integer id) {
+		if (userRepository.findById(id) == null) {
+			throw new BadRequestException(
+				new ErrorTO(
+					"http://github.com/adorogensky/workshop/alpha/user/id/not-found",
+					"Cannot delete user with id = '" + id + "'"
+				)
+			);
+		}
+
 		userRepository.deleteById(id);
 	}
 }
